requirements.txt
python-telegram-bot==21.10
fastapi==0.115.6
uvicorn==0.34.0
aiosqlite==0.19.0
python-dotenv==1.0.0
